Add convenience macro.
